Update tests and package names
